requirements.txt
flask>=2.0.0
Pillow>=9.0.0
PyAudio>=0.2.11
openai>=1.0.0
python-dotenv>=0.19.0
mss>=9.0.0
PyGetWindow>=0.0.9
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
chromadb>=0.4.0
pypdf>=3.0.0
sounddevice>=0.4.5
scipy>=1.9.0
numpy>=1.23.0
opencv-python>=4.6.0
tensorflow>=2.10.0
google-cloud-storage>=2.10.0